Add FileSystemHandle::move() method
Currently, it is not possible to move or rename a file or directory
without creating a new file/directory, copying over data (recursively,
in the case of a directory), and removing the original.

This CL allows for the atomic moving of a file or directory on the
local file system without needing to duplicate data.

Moves to non-local file systems will are not guaranteed to be atomic
and will involve duplicating data.

Bug: 1140805
Change-Id: I774ed1d9616249b6ecc80783db48a7bfee915aab